Address feedback

+++ typescript/utils/src/utils.ts
@@ -68,7 +68,7 @@ export const formatMessageV2 = (
   recipientAddr = addressToBytes32(recipientAddr);
 
   return ethers.utils.solidityPack(
-    ['uint32', 'uint256', 'uint32', 'bytes32', 'uint32', 'bytes32', 'bytes'],
+    ['uint8', 'uint256', 'uint32', 'bytes32', 'uint32', 'bytes32', 'bytes'],
     [
       version,
       nonce,
